move rope in glm out
